# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Spoonful",
    "version": "0.0.1",
    "description": "An API for managing and interacting with a database of recipes."
  },
  "servers": [
    {
      "url": "https://api.spoonful.com/v1"
    }
  ],
  "paths": {
...
    "/recipes/search": {
      "post": {
        "summary": "Search for recipes by by name or ingredients",
        "requestBody":{
          "content":{
            "application/json":{
              "schema":{
                "type":"object",
                "properties":{
                  "query":{
                    "type":"string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Recipe": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "ingredient": {
                  "type": "string"
                },
                "quantity": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}